+++ littletree/__init__.py
@@ -2,14 +2,12 @@ from abstracttree import MaxDepth
 
 from .basenode import BaseNode
 from .node import Node
-from .route import Route
 
 __all__ = [
     "BaseNode",
     "Node",
     "TreeMixin",
     "MaxDepth",
-    "Route",
 ]
 
 __version__ = "0.8.0"
