+++ setup.cfg
@@ -15,18 +15,16 @@ classifiers =
     Operating System :: OS Independent
     License :: OSI Approved :: Apache Software License
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.8
-    Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3.11
     Programming Language :: Python :: 3.12
     Typing :: Typed
 
 [options]
-python_requires = >= 3.8
+python_requires = >= 3.10
 packages = find:
 install_requires =
-    abstracttree >= 0.0.0
+    abstracttree >= 0.0.4
 
 [options.packages.find]
 include = littletree*
